chore: add base rule and consolidate common logic (#5849)
