+++ pkg/query-service/rules/rule.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"go.signoz.io/signoz/pkg/query-service/model"
+	v3 "go.signoz.io/signoz/pkg/query-service/model/v3"
 	"go.signoz.io/signoz/pkg/query-service/utils/labels"
 )
 
@@ -23,9 +24,8 @@ type Rule interface {
 
 	PreferredChannels() []string
 
-	Eval(context.Context, time.Time, *Queriers) (interface{}, error)
+	Eval(context.Context, time.Time) (interface{}, error)
 	String() string
-	// Query() string
 	SetLastError(error)
 	LastError() error
 	SetHealth(RuleHealth)
@@ -34,6 +34,8 @@ type Rule interface {
 	GetEvaluationDuration() time.Duration
 	SetEvaluationTimestamp(time.Time)
 	GetEvaluationTimestamp() time.Time
+
+	RecordRuleStateHistory(ctx context.Context, prevState, currentState model.AlertState, itemsToAdd []v3.RuleStateHistory) error
 
 	SendAlerts(ctx context.Context, ts time.Time, resendDelay time.Duration, interval time.Duration, notifyFunc NotifyFunc)
 }
